requirements.txt
openai
pydantic==1.10.18
noneprompt
pyyaml
httpx
beautifulsoup4
pytest
pytest-asyncio